Added the bar graph file back
